+++ test/addresses.unit.spec.js
@@ -125,5 +125,20 @@ describe('Test addresses', function() {
         assert.equal(addresses.addressToKey('Es3LFXNj5vHBw8c9kM98HKR69CJjUTyTPv4BdxoRbMQJ8zifxkgV').toString('hex'), '776b5cf08edea510711e2bc4a73f2b5118008906c5afd2e5786cf817fa279b80');
         assert.equal(addresses.addressToKey('Fs2w6VL6cwBqt6SpUyPLvdo9TK834gCr52Y225z8C5aHPAFav36X').toString('hex'), 'd48189215e445ea7e8dbf707c48922ab25a23552d8eae40cc5e9cd6b1a36963c');
     });
+
+    it('should generate random FCT address', function() {
+        const randomAddress = addresses.generateRandomFctAddress();
+        assert.isTrue(addresses.isValidFctPublicAddress(randomAddress.public));
+        assert.isTrue(addresses.isValidFctPrivateAddress(randomAddress.private));
+        assert.equal(addresses.getPublicAddress(randomAddress.private), randomAddress.public);
+    });
+
+    it('should generate random EC address', function() {
+        const randomAddress = addresses.generateRandomEcAddress();
+        assert.isTrue(addresses.isValidEcPublicAddress(randomAddress.public));
+        assert.isTrue(addresses.isValidEcPrivateAddress(randomAddress.private));
+        assert.equal(addresses.getPublicAddress(randomAddress.private), randomAddress.public);
+    });
+    
     
 });